% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baranov_f.R
\name{baranov_f}
\alias{baranov_f}
\title{Baranov F}
\usage{
baranov_f(F, C, B, M)
}
\arguments{
\item{F}{is the fishing mortality rate to solve.}

\item{C}{is a Numeric vector detailing the catch at \eqn{wk_{t}}}

\item{B}{is a Numeric vector of the biomass at \eqn{wk_{t}}}

\item{M}{is a Numeic vector of the natural mortality rate at \eqn{wk_{t}}}
}
\value{
Returns nothing, is objective to be solved by \code{\link{find_f}}
}
\description{
\code{baranov_f} provides the function to solve in
\code{\link{find_f}} for estimating weekly fishing mortality from catch
(\emph{C}), biomass (\emph{B}) and natural mortality (\emph{M}). It's based
on the standard Baranov catch equation.
}
\examples{
## No examples
}

