% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_areas.R
\name{close_areas}
\alias{close_areas}
\title{Close areas}
\usage{
close_areas(sim_init = sim_init, closure_init = NULL,
  commercial_logs = NULL, survey_logs = NULL, real_pop = NULL,
  t = t)
}
\arguments{
\item{closure_init}{is the output from \link{init_clousre}.}

\item{commercial_logs}{is the commercial landings data, the output from
\link{combine_logs}. Only needed if closure 'basis' = 'commercial'.}

\item{survey_logs}{is the survey data, the survey[["log.mat"]]. Only needed
if closure 'basis' is 'survey'.}

\item{real_pop}{is the popualtions as recorded. Only needed if closure
'basis' is 'real_pop'.}
}
\value{
is a list of closed cells, to pass to \link{go_fish}
}
\description{
The \code{close_areas} function implements the closures
according to the settings from \link{init_closure} and passes the areas to
\link{go_fish}. Its an internal function, requiring no user input.
}
\examples{
None
}
