% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_logs.R
\name{combine_logs}
\alias{combine_logs}
\title{Combine logs}
\usage{
combine_logs(fleets_catches)
}
\arguments{
\item{fleets_catches}{is the list output of fleets_catches from
\link{run_sim}}
}
\value{
is a dataframe of the fleet and vessel catches in logbook format
}
\description{
\code{combine_logs} is a helper function to convert the
list of fleet and vessels catch logs into a single dataframe.
}
\examples{
logs <- combine_logs(fleets_catches)
## Not run
}
