% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fields.R
\name{create_fields}
\alias{create_fields}
\title{Create species distribution fields}
\usage{
create_fields(
  npt = 1000,
  t = 1,
  seed = 123,
  n.spp = NULL,
  spp.ctrl = NULL,
  plot.dist = FALSE,
  plot.file = getwd()
)
}
\arguments{
\item{npt}{Numeric integer with the dimensions of the field in
\emph{npt * npt}}

\item{t}{Numeric integer with the number of time-steps in the simulation}

\item{seed}{(Optional) Numeric integer with the seed for the simulation}

\item{n.spp}{Numeric integer with the number of species to be simulated.
Each species must have an individual control list as detailed below.}

\item{spp.ctrl}{List of controls to generate each species spatio-temporal
distribution. Must be of the form spp.ctrl = list(spp.1 = c(rho0 = 0.001, ...),
spp.2 = c(rho0 = 0.001, ..),..) and contain the following:
\itemize{
    \item \strong{rho0} (\emph{>=0}) Controls the range in a matern covariance
    structure.
\item \strong{sigma2} (\emph{>=0}) Controls the marginal variance (i.e. process
error) in the matern (\emph{>=0}) covariance structure.
\item \strong{zeta} (\emph{>=0}) Damping parameter; regulates the temporal
correlation.
\item \strong{rho1} (\emph{>=0}) Range parameter for the diffusion process
\item \strong{gamma} (\emph{>=0}) Controls the level of anisotropy 
\item \strong{alpha} (\emph{[0, \eqn{\pi/2}]}) Controls the direction
of anisotropy
\item \strong{muX} (\emph{[-0.5, 0.5]}) x component of drift effect
\item \strong{muY} (\emph{[-0.5, 0.5]}) y component of drift effect
\item \strong{tau2} (\emph{>=0}) Nugget effect (measurement error)
\item \strong{nu} Smoothness parameter for the matern covariance
function
}}

\item{plot.dist}{Boolean, whether to plot the distributions to file}

\item{plot.file}{path to save the plots of the species distributions}
}
\value{
Silently returns a list of spatial distributions with first level of
the list being the population (1 -> n.spp) and the second being time (1 ->
t). If \code{plot.dist = TRUE} it produces an image of the spatial
distributions at each time step for each of the populations saved to the
working directory (unless specified otherwise in \code{plot.file})
}
\description{
\code{create_fields} parametrises and returns the
spatio-temporal fields used for the spatial distribution of fish populations
and movement in space and time for the simulations. 

The spatio-temporal fields are generated using
\code{\link[spate]{spate.sim}} function from the \emph{spate} package
using an advective-diffusion Stochastic Partial Differential Equation
(SPDE). See \emph{Lindgren 2011 and Sigrist 2015} for further detail.
}
\examples{
fields <- create_fields(n.spp = 1, t = 2, 
      spp.ctrl = list(
      'spp.1' = c('rho0' = 0.1, 'sigma2' = 1, 'zeta' = 0.1,
		  'rho1' = 0.01, 'gamma' = 0.3, 'alpha' = pi/4,
		  'muX' = -0.05, 'muY' = -0.05, 'tau2' = 0, 'nu' = 1.5)),
		plot.dist = TRUE, plot.file = getwd())
}
