% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hab.R
\name{create_hab}
\alias{create_hab}
\title{Create habitat distribution fields}
\usage{
create_hab(sim_init = sim, seed = 123, spp.ctrl = NULL,
  spawn_areas = NULL, spwn_mult = 10, plot.dist = FALSE,
  plot.file = getwd(), cores = 3)
}
\arguments{
\item{spp.ctrl}{List of controls to generate suitable habitat for each
species. Must be of the form spp.ctrl = list(spp.1 = c(var = 20, ...),
spp.2 = c(var = 10, ..),..) and contain the following:
\itemize{
\item \strong{nu} (\emph{>=0})
    \item \strong{var} (\emph{>=0}) Controls the range in a matern covariance
    \item \strong{scale} (\emph{>=0})
\item \strong{Aniso} (\emph{matrix, dim = c(2,2)})
}}

\item{plot.dist}{Boolean, whether to plot the distributions to file}

\item{plot.file}{path to save the plots of the species distributions}

\item{sim}{is the parameter settings for the simulation, made by
\code{init_sim} function.}
}
\value{
Silently returns a list of spatial distributions of suitable habitat
with first level of the list being the population (1 -> n.spp). If
\code{plot.dist = TRUE} it produces an image of the spatial distributions at
each time step for each of the populations saved to the working directory
(unless specified otherwise in \code{plot.file})
}
\description{
\code{create_hab} parametrises and returns the
spatial fields used for the distribution of suitable habitat for the
populations in the simulation. 

The spatial fields are generated using
\code{\link[RandomFields]{RFsimulate}} function from the \emph{RandomFields}
package.
}
\examples{
hab <- create_hab(sim.init = sim.init, spp.ctrl = list(
      'spp.1' = list('nu' = 1/0.15, var = 1, scale = 10, Aniso =
      matrix(nc=2, c(1.5, 3, -3, 4)))), spawn_areas = list("spp1" =
      list("area1" = c(2,4,6,8))), list("spp2" = list("area1" =
      c(0,10,23,35))), spwn_mult = 10, plot.dist = TRUE, plot.file =   getwd())
}
