% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spawn_hab.R
\name{create_spawn_hab}
\alias{create_spawn_hab}
\title{create spawning habitat}
\usage{
create_spawn_hab(hab = hab, spwnareas = NULL, mult = 10)
}
\arguments{
\item{hab}{is the habitat preference for the population}

\item{spwnareas}{is a list of Numeric vectors with the West, East, South and
North dimensions of the spawning areas, in the form list(spwn1 = c(x1, x2,
y1, y2)}

\item{mult}{is a Numeric with the attractiveness of the spawning area (a
multiplier)}
}
\value{
is the new habitat preference, taking account of the spawning area
}
\description{
\code{create_spawn_hab} modifies the habitat preference maps
created by \code{create_hab} to account for spawning habitat preference -
can be used as a substitute during spawning periods.
}
\examples{
create_spawn_hab(hab = matrix(nc = 100, runif(100 *
100)), spwnareas = list(spwn1 = c(20, 30, 50, 60)), mult = 10)

}
