% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_spawn.R
\name{define_spawn}
\alias{define_spawn}
\title{define spawning areas}
\usage{
define_spawn(coord = NULL, spwn = NULL, mult = 10)
}
\arguments{
\item{coord}{is a List of Numeric vectors of the boundaries of the spawning
areas, i.e. list(spwn1 = c(x1, x2, y1, y2), spwn2 = ...)}

\item{spwn}{is a Numeric matrix of 1s fed in by \code{create_spawn_hab}}

\item{mult}{is a Numeric of the attractiveness of the spawning areas}
}
\value{
a matrix of spawning preference
}
\description{
\code{define_spawn} is an auxiliary function called by
\code{create_spawn_hab} to create the spawning habitat preferences.
}
\examples{
define_spawn(coord = list(spwn1 = c(2,4,2,4)), spwn = matrix(nc = 3, runif(9)), mult = 10)  
}
