% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Degrees to radians}
\usage{
deg2rad(d)
}
\arguments{
\item{r}{is the bearing in radians}
}
\value{
is the bearing in degrees
}
\description{
\code{deg2rad} is a helper function to covert decimal degrees
to radians
}
\examples{
deg2rad(90)
}
