% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_difference.R
\name{delay_diff}
\alias{delay_diff}
\title{Delay-difference (weekly)}
\usage{
delay_diff(K = 0.3, F = NULL, M = 0.2, wt = 1, wtm1 = 0.1, R = NULL,
  B = NULL, Bm1 = NULL, al = NULL, alm1 = NULL)
}
\arguments{
\item{K}{is a Numeric vector describing growth.  Note:
K is transformed to rho with \emph{\eqn{\rho = exp{-K}}} for the model.
estimate of instantaneous fishing mortality (obtained elsewhere, via
\code{\link{find_f}} and \code{\link{baranov_f}} functions.}

\item{F}{is the weekly fishing mortality rate.}

\item{M}{is a Numeric vector of the instantaneous rate of natural mortality for the population}

\item{wt}{is a Numeric vector of the weight of a fish when fully recruited}

\item{wtm1}{is a Numeric vector of the weight of a fish before its recruited}

\item{R}{is a Numeric vector of the annual recruitment for the population in
numbers}

\item{B}{is the biomass of the population during \emph{\eqn{wk_{t}}}}

\item{Bm1}{is a Numeric vector of the biomass of the population in the previous week
\emph{\eqn{wk_{t-1}}}}

\item{al}{is a Numeric vector of the proportion of recruits to the fishery
in \emph{\eqn{wk_{t}}}}

\item{alm1}{is a Numeric vector of the proportion of recruits to the fishery
in \emph{\eqn{wk_{t-1}}}}
}
\value{
Returns the biomass at the beginning of the following week,
\emph{\eqn{wk_{t+1}}}
}
\description{
\code{delay_difference} implements a two-stage delay-difference
model with a weekly time-step after \emph{Dichmont 2003}. Given the starting
biomass, overall mortality and recruitment it returns the biomass in wk+1.
}
\examples{
delay_diff(K = 0.3, F = 0.2, M = 0.2, wt = 1, wtm1 = 0.1, R = 1e6, B = 1e5,
Bm1 = 1e4, al = 0.5, alm1 = 0.1)
}

