% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_calc}
\alias{distance_calc}
\title{distance calculation}
\usage{
distance_calc(x1, y1, x2, y2)
}
\arguments{
\item{x1}{is an integar for the starting x position}

\item{y1}{is an integar for the starting y position}

\item{x2}{is an integar for the end x position}

\item{y2}{is an integar for the end y position}
}
\value{
is a distance between the two cells
}
\description{
\code{distance_calc} calculates the euclidean distance between
two cell references.
}
\examples{
distance_calc(2, 3, 5, 7) 

}
