% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_calcR.R
\name{distance_calcR}
\alias{distance_calcR}
\title{Distance calc, R implementation (C++ seems slow??)}
\usage{
distance_calcR(x1 = 0, y1 = 0, x2 = 1, y2 = 1)
}
\arguments{
\item{x1}{}
}
\value{
distance
}
\description{
\code{distance_calcR} is a helper function to covert decimal degrees
to radians
}
\examples{
distance_calcR(x1 = 0, y1 = 0, x2 = 2, y2 = 9)
}
