% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_f.R
\name{find_f}
\alias{find_f}
\title{find F (fishing mortality)}
\usage{
find_f(C = C, B = B, M = M, FUN = baronov_f)
}
\arguments{
\item{C}{is a Numeric vector detailing the catch at \eqn{wk_{t}}}

\item{B}{is a Numeric vector of the biomass at \eqn{wk_{t}}}

\item{M}{is a Numeic vector of the natural mortality rate at \eqn{wk_{t}}}

\item{FUN}{is the objective function, here the Baranov equation
\link{baranov_f}}
}
\value{
Gives the fishing mortality estimate \emph{F}
}
\description{
\code{find_f} uses \code{\link[base]{uniroot}} to find the
fishing mortality rate given the catch, biomass and natural mortality using
the \code{\link{baranov_f}} objective function.
}
\examples{
find_f(C = 3000, B = 12000, M = 0.2, FUN = baranov_f)
}
