% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_spat_f.R
\name{find_spat_f}
\alias{find_spat_f}
\title{find spatial Fs (fishing mortality rates)}
\usage{
find_spat_f(sim_init = NULL, C = C, B = B, M = M, FUN = baranov_f)
}
\arguments{
\item{sim_init}{is the parameterised sim settings, made by \code{init_sim}}

\item{C}{is a Numeric vector detailing the catch at \eqn{wk_{t}}}

\item{B}{is a Numeric vector of the biomass at \eqn{wk_{t}}}

\item{M}{is a Numeic vector of the natural mortality rate at \eqn{wk_{t}}}

\item{FUN}{is the objective function, here the Baranov equation
\link{baranov_f}}
}
\value{
Gives a matrix the spatial fishing mortality estimate \emph{F}
}
\description{
\code{find_spat_f} uses \code{\link[base]{uniroot}} to find the
fishing mortality rate for a population given the catch, biomass and natural
mortality using the \code{\link{baranov_f}} objective function.
}
\examples{
find_spat_f(sim_init = sim, C = matrix(1000,3000, nc =2), B =
matrix(12000,10000, ncol = 2), M = 0.2, FUN = baranov_f)
}
