% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bearing.R
\name{get_bearing}
\alias{get_bearing}
\title{Get bearing function}
\usage{
get_bearing(b = NULL, k = NULL)
}
\arguments{
\item{b}{is a Numeric based on decimal degrees (0 - 360) of the current
bearing for the vessel}

\item{k}{is a Numeric [0-100] for the concentration parameter determining
the likely new direction for the vessel.}
}
\value{
bearing - is the new bearing for the vessel
}
\description{
\code{get_bearing} is a function to calculate a new bearing for
a vessel. The new bearing is determined from the Von Mises circular
distribution, with a concentration parameter, \emph{k} which is linked to
the value of the recent tow. Thus, if a vessel has a good tow, its more
likely to turn round and fish again in the same area.
}
\examples{
get_bearing(b = 270, k = 100)
}
