% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_fish.R
\name{go_fish}
\alias{go_fish}
\title{Go fish}
\usage{
go_fish(sim_init = sim, fleets_init = fleets, pops = Pop, t = t)
}
\arguments{
\item{sim_init}{is the initialised object from \code{init_sim}.}

\item{fleets_inits}{is the initialised object from \code{_init_fleets}}
}
\value{
is ...
}
\description{
\code{go_fish} is a function used to apply the fishing
simulation model
}
\examples{

}

