% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_fish_all_fleets.R
\name{go_fish_all_fleets}
\alias{go_fish_all_fleets}
\title{Go fish all fleets}
\usage{
go_fish_all_fleets(n_cores = 1, sim_init = sim, fleets = NULL,
  fleets_log = NULL, Pop = NULL, t = NULL)
}
\arguments{
\item{n_cores}{is the to use for the parallel processing}

\item{sim_init}{is the initialised sim object from \code{init_sim}}

\item{fleets}{is the initialised fleet object from \code{init_fleet}}

\item{fleets_log}{is the log of catches for fleets, containing a list of
fleets, each with a list of vessels, containing the vessel dataframe catches
and spatial catches}

\item{Pop}{is the current spatial populations biomass}

\item{t}{is the tow number}
}
\value{
is a list the same as fleets_log
}
\description{
\code{go_fish_all_fleets} applies the function
\code{go_fish_fleet} to each of the fleets with a parLapply using the
\code{parallel} package
}
\examples{
None as yet
}

