% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_closure.R
\name{init_closure}
\alias{init_closure}
\title{Initialise spatial closure(s)}
\usage{
init_closure(input_coords = NULL, basis = "commercial",
  rationale = "high_pop", spp1 = "spp1", spp2, year_start = 1,
  year_basis = NULL, closure_thresh = 0.95, sc = 1,
  temp_dyn = "annual")
}
\arguments{
\item{input_coords}{is a dataframe of x,y coordinates defining the
closure(s). If the temp_dyn are not static, the list should be multilayered
with the [[week/month]][x, y]}

\item{basis}{is a character string detailing the data used to define a
closure 'on the fly'. Can be \emph{survey} to be based on survey data,
\emph{commercial} to be based on commercial data, \emph{real_pop} to be
based on the simulated population. Not needed if coordinates defined.}

\item{rationale}{is the basis for any 'on the fly' closure. Can be
\emph{high_pop} for the areas of a highest population or \emph{high_ratio}
for the areas of the highest ratio of population 1: population 2. Not needed
if coordinates defined.}

\item{spp1}{is the first population as basis for the closure. If rationale =
high_pop then that should go here If rationale = high_ratio, its the target
(high quota) population. Not needed if coordinates are defined.}

\item{spp2}{is the second population when rationale = high_ratio, the lowest
quota population. Not needed if coordinates provided or rationale =
high_pop.}

\item{year_start}{is a Numeric indicating the first year the spatial
closure(s) shoud be implemented.}

\item{year_basis}{is a vector indicating the years of data the closure is
based on...Must be before \code{year_start}. If \code{NULL} then closure
will be calculated dynamically each year.}

\item{closure_thresh}{is the quantile of catches or high catch ratio which
determines closed cells}

\item{sc}{is a Numeric indicating the scale of data to use for the closure,
e.g. if the data is aggregated to 2 x 2 cells, is 2.}

\item{temp_dyn}{is a character string detailing whether closures should
be temporally 'annual', or change 'monthly' or 'weekly'.}
}
\value{
is a list of parameter settings for the spatial closures which
serves as an input to \link{run_sim}.
}
\description{
\code{init_closure} sets up the parameters for spatial
closure(s) in the simulation.
}
\examples{
Not as yet
}
