% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_fleet.R
\name{init_fleet}
\alias{init_fleet}
\title{Initialise fleet}
\usage{
init_fleet(
  sim_init = NULL,
  VPT = NULL,
  Qs = NULL,
  fuelC = 0,
  step_params = NULL,
  past_knowledge = FALSE,
  past_year_month = FALSE,
  past_trip = FALSE,
  threshold = NULL
)
}
\arguments{
\item{sim_init}{is the output (a list) from the \code{sim_init} function with the
indexing for the simulation.}

\item{VPT}{is a named vector of numerics detailing the value-per-tonne for catches
from each of the species (same for all fleets)}

\item{Qs}{is a list (an element for each fleet) with each element containing
a named vector with the catchability parameters for each species the vessels
in the fleet}

\item{fuelC}{is the fuel cost per unit of distance moved in euro}

\item{step_params}{is a list (an element for each fleet) with each element
containing a named vector with the step parameters used in
\code{step_length}. This must include the named elements \strong{rate},
\strong{B1}, \strong{B2}, \strong{B3}.}

\item{past_knowledge}{is a Boolean (TRUE / FALSE) whether past knowledge
should determine fishing location (only after the first year)}

\item{past_year_month}{is a Boolean (TRUE / FALSE) that indicates whether
the same month in previous years should be included in the past knowledge
decision}

\item{past_trip}{is a Boolean (TRUE / FALSE) that indicates whether the past
trip undertaken should be included in the past knowledge decision}

\item{knowledge_threshold}{is a numeric (0 - 1) detailing the threshold at
which a fishing tow should be considered "good" and included in the
selection of possible choices of starting fishing locations in future tows.}
}
\value{
is a list with three elements containing i) the fleet parameters,
a named list \strong{fleet_params}, ii) the fleet catches,
\strong{catches_list}, which is a list of a list. For
the\strong{catches_list} the first element denotes the fleet number, the
second element is the vessel number with a dataframe for recording the
vessels catches. Finally, iii) is the spatial catches for the fleets, which
is a list (fleet) containing a list (vessels) containing a list (population)
- which is to be passed to the delay difference model.
}
\description{
\code{init_fleet} sets up the parameters and results data frame
to record the catches from the simulation.
}
\examples{
None yet, to add
}
