% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_moveCov.R
\name{init_moveCov}
\alias{init_moveCov}
\title{Initialise movement covariates}
\usage{
init_moveCov(sim_init = NULL, steps = 52, spp_assoc = NULL)
}
\arguments{
\item{sim_init}{is the output from the function \link{init_sim}.}

\item{steps}{is a Numeric with the number of timesteps over which the
covariate changes}

\item{spp_assoc}{is a named list of whether species are positively
associated (+1) or negatively associated (-1) with the covariate.}
}
\description{
This function creates a list of covariates, to be used
}
\examples{
None
}
