% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pop.R
\name{init_pop}
\alias{init_pop}
\title{Initialise populations}
\usage{
init_pop(Bio = NULL, hab = NULL, start_cell = NULL, lambda = NULL,
  init_move_steps = 10)
}
\arguments{
\item{Bio}{is a Numeric vector of the starting (total) biomass for each of the
populations.}

\item{hab}{is the list of Matrices with the habitat preferences created by \code{create_hab}}

\item{start_cell}{is a list of Numeric vectors with the starting cells for
the populations}

\item{init_move_steps}{is a Numeric indicating the number of movements to
initialise for the population distributions}
}
\value{
The function returns the matrices for the starting spatial
population densities
}
\description{
\code{init_pop} sets up the populations spatial distribution
based on the habitat preference, starting cell and 'n' numbers of movements
for all populations in the simulation.
}
\examples{
init_pop(Bio = c(1e6, 2e5), hab = list(spp1 = matrix(nc = 10,
runif(10*10)), spp2 = matrix(nc = 10, runif(10*10)), init_move_steps = 10))
Note, example will not have the right biomass
}

