% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sim.R
\name{init_sim}
\alias{init_sim}
\title{Initialise simulation}
\usage{
init_sim(n_years = 1, n_tows_day = 4, n_days_wk_fished = 5,
  n_fleets = 1, n_vessels = 1, n_species = 1, move_length = 2)
}
\arguments{
\item{n_years}{is an integar defining the number of years for the
simulation}

\item{n_days_wk_fished}{is an integar defining the number of days in a
calendar week that are fished (e.g. 5 (out of 7))}

\item{n_fleets}{is an integar defining the number of fleets in the simulation}

\item{n_vessels}{is an integar defining the number of vessels in each fleet}

\item{n_species}{is an integar defining the number of species in the
simulation}

\item{move_length}{is an integar defining the duration (in weeks) between
spatial movements for the populations}

\item{n_tow_day}{is an integar defining the number of tows in a days fishing}
}
\value{
is a list of lists, detailing the indexs and data formats necessary
for the simulation.
}
\description{
\code{init_sim} sets up the general simulation parameters such
as number of tows in a day, number of days fished in a week, how often
species movement occurs and number of years for the simulation. It also
creates some vector and matrix structures which are used in the
\code{init_pop} and \code{init_fleet} functions.
}
\examples{
init_sim(n_years = 1, n_tows_day = 4, n_days_wk_fished = 5,
n_fleets = 1, n_vessels = 1, n_species = 1, move_length = 2)
}

