% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_survey.R
\name{init_survey}
\alias{init_survey}
\title{Initialise survey settings}
\usage{
init_survey(sim_init = NULL, design = "fixed_station",
  n_stations = 50, start_day = 90, stations_per_day = 5, Qs = NULL)
}
\arguments{
\item{sim_init}{is the general simualtion settings from \link{sim_init}}

\item{design}{is the survey design used, at the moment only
\emph{fixed_station}}

\item{n_stations}{is a Numeric for the number of stations to be fished each.
Note: If using 'fixed_station' design this will be rounded down to maintain
a grid shape if not divisble.}

\item{start_day}{is a Numeric for the first day of the survey each year}

\item{stations_per_day}{is a Numeric for the number of stations surveyed per
day}

\item{Qs}{is a named Numeric Vector containing any survey catchabilities, assumed to be time invariant.}
}
\value{
is a list consisting of the survey setting and a a matrix for
storing the log of catches from the survey, to be used as an input to
\link{run_sim}.
}
\description{
\code{init_survey} is a function to mimic a
fisheries-independent survey to sample catches from the populations.
}
\examples{
init_survey(design = 'fixed_station', n_stations = 50, start_day = 90, stations_per_days = 5, Qs = c("spp1" = 0.1, "spp2" = 0.2)
}
