% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic probability}
\usage{
logistic(A = 0, K = 0.95, C = 1, Q = 200, B = NULL, v = 1, t = NULL)
}
\arguments{
\item{A}{is the lower asymptote, set at 0}

\item{K}{is the upper asymptote, set at 0.95 (to keep some few exploratory
tows, even when established)}

\item{C}{= 1}

\item{Q}{defines the lower curve, related to Y at 0, usefully set at
tmax/100}

\item{B}{is the growth rate, e.g· 0.001}

\item{v}{affects bear where asymptote maximum growth occurs, set at 1}
}
\description{
\code{logistic} is a helper function to generate a logistic
curve to transition through the fishery stages (exploratory > transition >
established). Where ~0 is exploratory fishing and ~1 is established on past
knowledge. 
Only Q (can be set at tmax/100) and t (current tow) need to be supplied.
}
\examples{
 NOT RUN
}
