% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_step.R
\name{make_step}
\alias{make_step}
\title{make step function}
\usage{
make_step(stepD, Bear, start.x, start.y)
}
\arguments{
\item{stepD}{is a Numeric vector of the distance to move}

\item{Bear}{is a Numeric vector of the bearing to move (in degrees)}

\item{start.x}{is the starting point on the x-axis}

\item{start.y}{is the starting point on the y-axis}
}
\value{
returns a new coordinate position through a vector (x, y)
}
\description{
\code{make_step} determines the new position of the vessel
following a move, using the step distance and bearing inputs.
}
\examples{
make_step(stepD = 20, Bear = 90, start.x = 20, start.y = 5)
}

