% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_population}
\alias{move_population}
\title{population movement function}
\usage{
move_population(moveProp, StartPop)
}
\arguments{
\item{moveProp}{is a list of the proportion of the population from each
cell to reallocated to each of the other cells}

\item{StartPop}{is a Numeric Matrix of the current populations distribution}
}
\value{
is a list of the new position for the population from each of the
cells. 

NOTE: This is not aggregated and requires calling the R function
\emph{Reduce('+', Lst)} to reaggregate. Would be better if done in function
but Reduce is currently faster...but much more memory intensive to get out
the lists...using the standard c++ accumulate
function may work for this but untested
}
\description{
\code{move_population} redistributes the population based on
the movement probabilities
}
\examples{
None at the moment

}

