% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_prob}
\alias{move_prob}
\title{movement probability function}
\usage{
move_prob(start, lambda, hab)
}
\arguments{
\item{start}{is a Numeric vector of dim 2 for the starting position c(x,y)}

\item{lambda}{is an integar for the value for the exponential decay in probability 
of movement, i.e. \eqn{Pr(B|A) = \exp{-\lambda*dist_{a,b}} / Sum(c=1:c=n)  \exp{-\lambda*dist}}}

\item{hab}{is a matrix of the habitat suitability}
}
\value{
is a matrix of the movement probabilities from a cell
}
\description{
\code{move_prob} calculates the movement probability between 
a cell and all other cells based on the distance and \emph{lambda}.
}
\examples{
move_prob(c(2, 5), 0.3, matrix(nc = 3, runif(9))) 

}
