% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_prob_Lst}
\alias{move_prob_Lst}
\title{movement probability function as a list}
\usage{
move_prob_Lst(lambda, hab)
}
\arguments{
\item{lambda}{is the decay value as in \code{move_prob}}

\item{hab}{is a matrix of the habitat suitability for the population}
}
\value{
is a list of the movement probabilities form each cell to all other
cells
}
\description{
\code{move_prob_list} applies \code{\link{move_prob}} from all
cells to all other cells and returns as a list.
}
\examples{
None at the moment

}

