% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_dist.R
\name{norm_fun}
\alias{norm_fun}
\title{Normal distribution}
\usage{
norm_fun(x, mu, va)
}
\arguments{
\item{move_init}{is the output from \link{init_moveCov}}
}
\description{
Helper function used for returning the PDF of a normal
distribution from the supplied temperature tolerances in \link{init_moveCov}
}
\examples{
sapply(seq(2,20,0.1), mu = 10, va = 6)
}
