% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_comp.R
\name{plot_catch_comp}
\alias{plot_catch_comp}
\title{Plot the spatial catch composition from the commercial catches as
'square pie charts' using \link{mapplots}.}
\usage{
plot_catch_comp(gran = c(20, 10, 5), logs = logs, fleets = 1:2,
  vessels = 1:5, trips = 1:20, years = 18:20, cluster_plot = FALSE,
  cluster_k = 5)
}
\arguments{
\item{gran}{is a Numeric Vector of granularities required}

\item{logs}{is the fleet logs from \link{combine_logs}}

\item{fleets}{is a Numeric Vector of the fleets to include in the catch
composition plot}

\item{vessels}{is a Numeric Vector of the vessels to include in the plot}

\item{trips}{is a Numeric Vector of the trips to include}

\item{years}{is a Numeric Vector of the years}

\item{cluster_plot}{is a logical, determines whether also to run PAM cluserting
on  the catch compositions and plot the clusters spatially}

\item{clusters_k}{is the number of clusters to search for in the PAM
clustering algorithm}
}
\description{
Plotting of spatial catch compositions at different levels of
aggregation
}
\examples{
plot_catch_comp(gran = c(20,10,5,2), logs = logs, fleets = 1:2, vessels =
1:5. trips = 1:20, years = 18:20, cluster_plot = FALSE, cluster_k = 5)
}
