% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily_fdyn.R
\name{plot_daily_fdyn}
\alias{plot_daily_fdyn}
\title{Plot daily fishing mortality dynamics}
\usage{
plot_daily_fdyn(results)
}
\arguments{
\item{results}{is output from the function \link{run_sim}.}
}
\value{
is a matplot of the daily fishing mortality dynamics
}
\description{
\code{plot_daily_fdyn} plots the daily fishing mortality
dynamics by year.
}
\examples{
plot_daily_fdyn(results = results)
}
