% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fleet_trip.R
\name{plot_fleet_trip}
\alias{plot_fleet_trip}
\title{Plot an entire fleet for a trip}
\usage{
plot_fleet_trip(
  logs = logs,
  fleet_no = 1,
  year_trip = 1,
  trip_no = 1,
  pop_bios = NULL,
  fleets_init = NULL,
  sim_init = NULL
)
}
\arguments{
\item{logs}{is the combined log file, from \link{combine_logs}.}

\item{fleet_no}{is a Numeric, the fleet from which to plot}

\item{year_trip}{is a Numeric, the year in which the trip took place}

\item{trip_no}{is a Numeric for the trip you wish to plot}
}
\description{
\code{plot_fleet_trip} is a plot of a whole fleets vessels movement
during one trip. It's intended for diagnostics.
}
\examples{
plot_fleet_trip(logs = logs, fleet_no = 1, year_trip = 1, trip_no = 1,
pop_bios = NULL, fleets_init = NULL)
}
