% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_habitat.R
\name{plot_habitat}
\alias{plot_habitat}
\title{Plot the habitat distribution fields for all populations}
\usage{
plot_habitat(hab = hab)
}
\arguments{
\item{hab}{is the habitat to plot, either unadjusted or adjusted for
spawning area.}
}
\value{
Silently plots the habitat fields (can be either the unadjusted
habitat, or including the spawning adjustment).
}
\description{
\code{plot_habitat} is a function to plot the habitat spatial
fields.
}
\examples{
plot_habitat(hab = hab)
}
