% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop_summary.R
\name{plot_pop_summary}
\alias{plot_pop_summary}
\title{Plot population summary}
\usage{
plot_pop_summary(results = res, timestep = "weekly", save = FALSE,
  save.location = ".")
}
\arguments{
\item{results}{is an output from the function \link{run_sim}.}

\item{timestep}{is a character string determining whether the plot is
'weekly' or 'annual'}

\item{save}{is a logical whether to save the plot}

\item{save.location}{is a location (defaults to current directory)}
}
\value{
is a ggplot of all the species and metrics as a faceted plot
examples
plot_pop_summary(results = res, timestep = 'weekly', save = TRUE, location = '.') 
## Not run
}
\description{
\code{plot_pop_summary} plots the four population dynamic
metrics: catches, biomass, fishing mortality and recruitment.
It can either operate at a weekly timestep or an annual timestep
}
