% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_realised_stepF.R
\name{plot_realised_stepF}
\alias{plot_realised_stepF}
\title{Plot realised step function}
\usage{
plot_realised_stepF(logs = logs, fleet_no = 1, vessel_no = 1)
}
\arguments{
\item{logs}{is the log file from \link{combine_logs}}

\item{fleet_no}{is a Numeric of the fleet to plot}

\item{vessel_no}{is a Numeric of the vessel to plot}
}
\description{
\code{plot_realised_stepF} diagnostics plot of the step function shape realised in the
simulation
}
\examples{
plot_realised_stepF(logs = logs, fleet_no = 1, vessel_no = 1)
}
