% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatiotemp_hab.R
\name{plot_spatiotemp_hab}
\alias{plot_spatiotemp_hab}
\title{Plot spatiotemporal habitat suitability}
\usage{
plot_spatiotemp_hab(hab = NULL, moveCov = NULL, plot.file = getwd(),
  spwn_wk = NULL)
}
\arguments{
\item{hab}{is the output from \link{create_hab}}

\item{moveCov}{is the output from \link{init_moveCov}}

\item{plot.file}{path to save the plots of the spatiotemporal habitats}

\item{spwn_wk}{is a named list of the spawning week for each population}
}
\description{
Function to plot out the habitat suitability, as adjusted by
the spatiotemporal move covariates
}
\examples{
None
}
