% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survey.R
\name{plot_survey}
\alias{plot_survey}
\title{Plot the fisheries independent survey results}
\usage{
plot_survey(survey = NULL, type = "spatial")
}
\arguments{
\item{survey}{is the survey results from the \link{run_sim} function.}

\item{type}{is a character indicating if \emph{spatial} or \emph{index}}
}
\value{
is a plot of the spatial distribution of survey catches and an
inter-annual abundance index
}
\description{
\code{plot_survey} plots the spatial abundances and an index
from the fisheries independent survey, for each population.
}
\examples{
plot_survey(survey = survey, type = "spatial")
}
