% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2.deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to degrees}
\usage{
rad2deg(r)
}
\arguments{
\item{d}{is the bearing in decimal degrees}
}
\value{
is the bearing in radians
}
\description{
\code{rad2deg} is a helper function to covert radians to decimal degrees
}
\examples{
rad2deg(2)
}
