% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recr.R
\name{Recr}
\alias{Recr}
\title{Recruitment function}
\usage{
Recr(model, params, B, cv, ..)
}
\arguments{
\item{model}{is a character detailing the recruitment function to use
(currently 'BH' for Beverton and Holt or 'Ricker' for a Ricker stock-recruit
relationship.}

\item{params}{is a Numeric vector of length 2, containing labelled \emph{a} and
\emph{b} parameters for the stock-recruit function. 
For Beverton and Holt \emph{a} refers to the maximum recruitment rate in
biomass, \emph{b} refers to the Spawning Stock Biomass (SSB) required to
produce half the maximum.
For Ricker \emph{a} refers to the maximum productivity per spawner and
\emph{b} the density dependent reduction in productivity as the stock
increases.}

\item{B}{is a Numeric vector containing the SSB of the adult population from
which the recruitment derives.}

\item{cv}{is a Numeric vector containing the coefficient of variation in the
recruitment function.}
}
\value{
returns the recruitment to the population in biomass.
}
\description{
\code{Recr} returns a biomass of recruited fish to the
population based on a stock-recruit relationship and some measure of
variation.
}
\examples{
Recr(model = 'BH', params = c("a" = 2000, "b" = 200), B = 1000, cv = 0.1)
}
