% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run sim}
\usage{
run_sim(sim_init = NULL, pop_init = NULL, move_cov = NULL,
  fleets_init = NULL, hab_init = NULL, InParallel = TRUE,
  save_pop_bio = FALSE, survey = NULL, closure = NULL, ...)
}
\arguments{
\item{sim_init}{is the parameterised simulation settings from
\code{init_sim}}

\item{pop_init}{is the parameterised populations from \code{init_pop}}

\item{move_cov}{is a parameterised movement covariate object, from
\code{init_moveCov}}

\item{fleets_init}{is the parameterised fleets from \code{init_fleets}}

\item{hab_init}{is the parameterised habitat maps from \code{create_hab}}

\item{InParallel}{is a BOLEEN indicating whether calculations should be done
using parallel processing from \code{parallel}, default is TRUE}

\item{save_pop_bio}{is a logical flag to indicate if you want to record #' true spatial population at each time step (day)}

\item{survey}{is the survey settings from \link{init_survey}, else NULL if no survey is due to be simulated}

\item{closure}{is the spatial closure settings from \link{init_closure}m
else NULL if no closures are to be implemented}
}
\value{
is the results...
}
\description{
\code{run_sim} is the overarching simulation function, taking
all the parameterised inputs and returning the results.
}
\examples{
Not yet

}
