% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_length.R
\name{step_length}
\alias{step_length}
\title{Step length function}
\usage{
step_length(step_params = params[["step_params"]], revenue = revenue)
}
\arguments{
\item{step_params}{is a list of parameters which determine the relationship between
revenue gained from the recent fishing activity and the next move step
length, based on a gamma function. The list contains the following:
\itemize{
\item \strong{rate} Determines the rate ....
\item \strong{B1} Determines...
\item \strong{B2} Determines ...
\item \strong{B3} Determines ..
}}

\item{revenue}{is the last observed fishing revenue for the vessel}
}
\value{
step - the size of the next step
}
\description{
\code{step_length} is a function to calculate the step length a
vessel takes based on the step parameters provided for a gamma function and
the revenue from the most recent fishing activity.
}
\examples{
step_length(step_params = list(B1 = 1, B2 = 50, B3 = 2000, rate = 1),
revenue = 300)
}
