% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_fleet_catches.R
\name{sum_fleet_catches}
\alias{sum_fleet_catches}
\title{Sum fleet catches}
\usage{
sum_fleet_catches(fleet_log = NULL, n_spp = NULL)
}
\arguments{
\item{fleet_log}{is the output of \code{go_fish_fleet}, i.e. the catch log
information for a single fleet}

\item{n_spp}{is the number of populations in the simulation (NOTE: can
remove this and take from the overall sim settings)}
}
\value{
is a list of matrices (one for each population) with the entire
fleets catches of the population
}
\description{
\code{sum_fleet_catches} is a helper function to take the
spatial catches for an entire fleet and sum them as a matrix of catches for
the fleet for each population
}
\examples{
test <- sum_fleet_catches(fleet_log = applied_to_fleets[[1]], n_spp = 2)
}

