% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_fleets_catches.R
\name{sum_fleets_catches}
\alias{sum_fleets_catches}
\title{Sum fleets catches}
\usage{
sum_fleets_catches(FUN = sum_fleet_catches, fleets_log = NULL,
  sim_init = sim, ...)
}
\arguments{
\item{FUN}{is the function, i.e. \code{sum_fleet_catches}}

\item{fleets_log}{is the log of all the catches for all fleets, coming from
application of go_fish_fleet to all fleets}

\item{n_spp}{is the number of populations in the simulation (NOTE: can
remove this and take from the overall sim settings)}
}
\value{
is a list of matrices (one for each population) with all fleets
catches of each population. This is then used as an input to the baranov
calcs
}
\description{
\code{sum_fleets_catches} is a helper function to apply
\code{sum_fleet_catches} to all fleets, returning a single list of matrices
with the catches of each population across all fleets and vessels.
}
\examples{
spp_catches <- sum_fleets_catches(FUN = sum_fleet_catches,
fleets_log = applied_to_fleets, n_spp = 2)
}
