% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_step.R
\name{test_step}
\alias{test_step}
\title{test step length function}
\usage{
test_step(step_params = step_params, rev.max = 2000)
}
\arguments{
\item{step_params}{is a list of parameters which determine the relationship between
revenue gained from the recent fishing activity and the next move step
length, based on a gamma function. The list contains the following:
\itemize{
\item \strong{rate} Determines the rate ....
\item \strong{B1} Determines...
\item \strong{B2} Determines ...
\item \strong{B3} Determines ..
}}

\item{rev.max}{is the maximum revenue at which to test the step length
function.}
}
\value{
is a plot of the relationship between revenue and step length
}
\description{
\code{test_step} is a function to test and review parameters
for the step_length function. This is primarily to help with identifying the
right parameters for the desired relationship between revenue and step
length.
}
\examples{
test_step(step_params = list(B1 = 1, B2 = 50, B3 = 2000, rate = 1), rev.max
= 2000)
}
