% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_past_knowledge.R
\name{use_past_knowledge}
\alias{use_past_knowledge}
\title{Use past knowledge}
\usage{
use_past_knowledge(p = NULL)
}
\arguments{
\item{p}{is the probability of using past knowledge, drawn from
\link{logistic}}
}
\description{
\code{use_past_knowledge} is a helper function to make a
random draw whether to do exploratory fishing or go to known fishing grounds
}
\examples{
None
}
