% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hab.R
\name{create_hab}
\alias{create_hab}
\title{Create habitat distribution fields}
\usage{
create_hab(npt = 100, seed = 123, n.spp = NULL, spp.ctrl = NULL,
  plot.dist = FALSE, plot.file = getwd())
}
\arguments{
\item{npt}{Numeric integer with the dimensions of the field in
\emph{npt * npt}}

\item{n.spp}{Numeric integer with the number of species to be simulated.
Each species must have an individual control list as detailed below.}

\item{spp.ctrl}{List of controls to generate suitable habitat for each
species. Must be of the form spp.ctrl = list(spp.1 = c(var = 20, ...),
spp.2 = c(var = 10, ..),..) and contain the following:
\itemize{
\item \strong{nu} (\emph{>=0})
    \item \strong{var} (\emph{>=0}) Controls the range in a matern covariance
    \item \strong{scale} (\emph{>=0})
\item \strong{Aniso} (\emph{matrix, dim = c(2,2)})
}}

\item{plot.dist}{Boolean, whether to plot the distributions to file}

\item{plot.file}{path to save the plots of the species distributions}
}
\value{
Silently returns a list of spatial distributions of suitable habitat
with first level of the list being the population (1 -> n.spp). If
\code{plot.dist = TRUE} it produces an image of the spatial distributions at
each time step for each of the populations saved to the working directory
(unless specified otherwise in \code{plot.file})
}
\description{
\code{create_hab} parametrises and returns the
spatial fields used for the distribution of suitable habitat for the
populations in the simulation. 

The spatial fields are generated using
\code{\link[RandomFields]{RFsimulate}} function from the \emph{RandomFields}
package.
}
\examples{
fields <- create_hab(n.spp = 1,  
      spp.ctrl = list(
      'spp.1' = list('nu' = 1/0.15, var = 1, scale = 10, Aniso =
      matrix(nc=2, c(1.5, 3, -3, 4)))), plot.dist = TRUE, plot.file =
      getwd())
}

