% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_iter.R
\name{fish_iters}
\alias{fish_iters}
\title{Go fish iter}
\usage{
fish_iters(FUN, N = 1, fleets = NULL, Pop = NULL, t = t, ...)
}
\arguments{
\item{fleets}{is the parameter settings and catch recording dataframe for all fleets}

\item{Pop}{is the population matrix for all populations}
}
\value{
is a list with the objects catch detailing the fleet catches and
catch_matrix detailing the spatial catches, to input to the delay difference
model
}
\description{
\code{go_fish_iter} applies the function \code{go_fish} to the
entire fleet an lapply.
}
\examples{
None as yet
}

