% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_fish.R
\name{go_fish}
\alias{go_fish}
\title{Go fish}
\usage{
go_fish(sim_init = NULL, fleet_params = NULL, fleet_catches = NULL,
  sp_fleet_catches = NULL, pops = NULL, closed_areas = NULL, t = t)
}
\arguments{
\item{sim_init}{is the initialised object from \code{init_sim}.}

\item{fleet_params}{is the parameter settings initialised from \code{_init_fleets}}

\item{fleet_catches}{is the DF initialised from \code{_init_fleets}}

\item{sp_fleet_catches}{is a list of spatial catches (as a Numeric matrix) for the fleet of each population}

\item{closed_areas}{is a dataframe with the x,y coordinates are any closed
areas, provided internally by \link{close_areas}}
}
\value{
is a list containing i) the fleet catch dataframes , ii) the spatial
catches of each population
}
\description{
\code{go_fish} is a function used to apply the fishing
simulation model
}
\examples{

}
