% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_moveCov.R
\name{init_moveCov}
\alias{init_moveCov}
\title{Initialise movement covariates}
\usage{
init_moveCov(sim_init = NULL, steps = 52, spp_tol = NULL)
}
\arguments{
\item{sim_init}{is the output from the function \link{init_sim}.}

\item{steps}{is a Numeric with the number of timesteps over which the
covariate changes}

\item{spp_tol}{is a named list (each species) with a list of mean (mu) and
variance (va) for the normal distribution for thermal tolerance.}
}
\description{
This function creates a list of covariates, to be used
}
\examples{
None
}
