% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vessel_move.R
\name{plot_vessel_move}
\alias{plot_vessel_move}
\title{Plot vessel move}
\usage{
plot_vessel_move(logs = logs, fleet_no = 1, vessel_no = 1,
  year_trip = 1, trip_no = 1)
}
\arguments{
\item{logs}{is the combined log file, from \link{combine_logs}.}

\item{fleet_no}{is a Numeric, the fleet from which to plot}

\item{vessel_no}{is a Numeric, the vessel to plot from the chosen fleet}

\item{year_trip}{is a Numeric, the year in which the trip took place}

\item{trip_no}{is a Numeric for the trip you wish to plot}
}
\description{
\code{plot_vessel_move} is a plot of a single vessel movement
during one trip. It's intended for diagnostics.
}
\examples{
plot_vessel_move(logs = logs, fleet_no = 1, vessel_no = 1, year_trip = 1, trip_no = 1)
}
