% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to degrees}
\usage{
rad2deg(r)
}
\arguments{
\item{r}{is the bearing in radians}

\item{d}{is the bearing in degrees}
}
\description{
\code{rad2deg} is a helper function to covert radians to
decimal degrees
}
\examples{
rad2deg(pi)
}
