% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run sim}
\usage{
run_sim(sim_init = NULL, pop_init = NULL, fleets_init = NULL,
  hab_init = NULL, InParallel = TRUE, cores = 3, ...)
}
\arguments{
\item{sim_init}{is the parameterised simulation settings from
\code{init_sim}}

\item{pop_init}{is the parameterised populations from \code{init_pop}}

\item{fleets_init}{is the parameterised fleets from \code{init_fleets}}

\item{hab_init}{is the parameterised habitat maps from \code{create_hab}}

\item{InParallel}{is a BOLEEN indicating whether calculations should be done
using parallel processing from \code{parallel}, default is TRUE}
}
\value{
is the results...
}
\description{
\code{run_sim} is the overarching simulation function, taking
all the parameterised inputs and returning the results.
}
\examples{
Not yet

}
