% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_fleet_catches.R
\name{sum_fleet_catches}
\alias{sum_fleet_catches}
\title{Sum fleet catches}
\usage{
sum_fleet_catches(sim_init = sim, fleet_log = NULL)
}
\arguments{
\item{sim_init}{is the initialised simulation settings, from \code{init_sim}}

\item{fleet_log}{is the output of \code{go_fish_fleet}, i.e. the catch log
information for a single fleet}
}
\value{
is a list of matrices (one for each population) with the entire
fleets catches of the population
}
\description{
\code{sum_fleet_catches} is a helper function to take the
spatial catches for an entire fleet and sum them as a matrix of catches for
the fleet for each population
}
\examples{
test <- sum_fleet_catches(fleet_log = applied_to_fleets[[1]])
}
